src/mainwindow.cpp: don't draw on screen until ready
The OS signal handler can't be installed until after the Qt event loop
starts since it eventually calls `QApplication::exit(0);`, and that will
cleanup anything we draw on the screen. Thus we don't want to connect
any slots that touch the what's in focus or draw on the screen until the
OS signal handler is ready. `mainwindow->init()` is called in
`src/main.cpp` before `app.exec()`, thus we don't want to connect any
slots that draw on screen in `mainwindow->init()`. Instead,
`mainwindow->init()` will connect the `setupWindow()` slot, to the the
`signalHandlerInstalled` Qt signal emitted by `this->handler`. At that
point the OS signal handler has been installed and it is safe to draw on
the screen.